+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## dev-develop
 
+ - FEATURE #68 Added checkboxes to deactivate notify and success emails
  - FEATURE #67 Added sixths widths to fields
  - BUGFIX #66 Fixed that dynamic is related to a form
  - BUGFIX #65 Unlimited title column and changed string to text